requirements.txt
streamlit==1.35.0
pillow==10.3.0
google-generativeai~=0.7
python-dotenv~=1.0.1
google-cloud-texttospeech==2.16.3
gTTS~=2.5.1
playsound~=1.3.0